Fix relation bugs && Add Patch to ChatMember

+++ backend/src/chat/membership.service.ts
@@ -4,6 +4,7 @@ import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateChatMemberDto } from './dto/create-chat-member.dto';
 import { UserService } from 'src/user/user.service';
 import { ChatService } from './chat.service';
+import { ChatRoleDto } from './dto/update-chat-member.dto';
 
 @Injectable()
 export class MembershipService {
@@ -25,7 +26,7 @@ export class MembershipService {
     const chat = await this.chatService.findChatById(chatId)
     const user = await this.userService.findUserById(createChatMemberDto.id)
     if(!chat || !user)
-    throw new NotFoundException(`${'Chat' ? !chat : 'User'} not found`);
+      throw new NotFoundException(`${'Chat' ? !chat : 'User'} not found`);
     await this.checkAccess(chat, createChatMemberDto);
     try {
       await this.prisma.chatMember.create({
@@ -72,15 +73,15 @@ export class MembershipService {
     } catch(error) {
       throw new ForbiddenException('Could not delete chat member')
     }
-    if(!newOwner) {
+    if(chatMember.owner === true && !newOwner) {
       try {
         this.chatService.deleteChat(chatId);
       } catch(error) {
         throw new ForbiddenException('Could not delete chat');
       }
     }
-    else
-      await this.updateChatMember(newOwner, chatId, {owner: true, administrator: true});
+    else if(chatMember.owner === true)
+      await this.updateChatMember(newOwner.userId, chatId, {owner: true, administrator: true});
   }
 
   private async findNewOwner(userId: number) {
@@ -90,22 +91,45 @@ export class MembershipService {
           not: userId
         },
       },
-      orderBy: {
-        administrator: 'desc',
-        createdAt: 'asc'
+      orderBy: [
+        {
+          administrator: 'desc'
+        },
+        {
+          createdAt: 'asc'
+        }
+      ]
       }
-    });
+    );
   }
 
-  private async updateChatMember(newOwner: any, chatId: number, newData: object) {
-    await this.prisma.chatMember.update({
-      where: {
-        chatId_userId: {
-          userId: newOwner.userId,
-          chatId: chatId
+  async updateChatMember(userId: number, chatId: number, newData: ChatRoleDto) {
+    const user = await this.findChatMemberByIds(chatId, userId);
+    if(!user)
+      throw new NotFoundException('Chat member not found')
+    if(newData.owner === true) {
+      newData.administrator = true;
+      const owner = await this.prisma.chatMember.findFirst({
+        where: {
+          chatId: chatId,
+          owner: true
         }
-      },
-      data: newData
-    });
+      })
+      if(owner)
+        await this.updateChatMember(owner.userId, chatId, {administrator: true, owner: false})
+    }
+    try {
+      await this.prisma.chatMember.update({
+        where: {
+          chatId_userId: {
+            userId: userId,
+            chatId: chatId
+          }
+        },
+        data: newData
+      });
+    } catch(error) {
+      throw new ForbiddenException('Could not update user');
+    }
   }
 }
